add test manufacturer images to gitignore. Fix specificPriceError when price or reduction not provided & customer search error

+++ src/Adapter/Product/SpecificPrice/CommandHandler/EditProductSpecificPriceHandler.php
@@ -85,7 +85,7 @@ class EditProductSpecificPriceHandler implements EditProductSpecificPriceHandler
         }
 
         if ($command->getPrice()) {
-            $specificPrice->price = (float) (string) $command->getPrice();
+            $specificPrice->price = (string) $command->getPrice();
             $updatableProperties[] = 'price';
         }
